Added User-Defined Syntax Delimiters
End-users can define the syntax tokens used as the frontend for templating.
This allows end-users to prevent template collision when using a template
to create a template for a different templating syntax. Example of ERB-like
templating syntax included in tests. Additionally reordered out / out-unescaped
DCG rules to favor syntaxes which include the "unescape" keyword but use the
same template start tag. Both existing syntaxes are succinctly defined in
the same manner as a user defined syntax.